Templatize class to minimize code size.
The MicroOLED class is now three different classes for the SPI, I2C, and
Parallel communication modes, with the functionality in a templatized
common base class. This avoids always building in the I2C, SPI, and
Parallel comm code even though only one will likely be used.

When using SPI mode, this resulted in a hex file using 1092 bytes less,
not insignificant.